+++ backend/windmill-worker/src/worker.rs
@@ -20,6 +20,7 @@ use std::{
     collections::hash_map::DefaultHasher,
     hash::{Hasher, Hash},
 };
+use regex::Regex;
 
 use tracing::{trace_span, Instrument};
 use uuid::Uuid;
@@ -43,9 +44,6 @@ use tokio::{
     },
     time::{interval, sleep, Instant, MissedTickBehavior}
 };
-
-#[cfg(feature = "enterprise")]
-use tokio::join;
 
 use futures::{
     future::{self, ready, FutureExt},
@@ -135,17 +133,22 @@ pub const ROOT_TMP_CACHE_DIR: &str = "/tmp/windmill/tmpcache/";
 pub const LOCK_CACHE_DIR: &str = concatcp!(ROOT_CACHE_DIR, "lock");
 pub const PIP_CACHE_DIR: &str = concatcp!(ROOT_CACHE_DIR, "pip");
 pub const DENO_CACHE_DIR: &str = concatcp!(ROOT_CACHE_DIR, "deno");
+pub const DENO_CACHE_DIR_DEPS: &str = concatcp!(ROOT_CACHE_DIR, "deno/deps");
+pub const DENO_CACHE_DIR_NPM: &str = concatcp!(ROOT_CACHE_DIR, "deno/npm");
+
 pub const GO_CACHE_DIR: &str = concatcp!(ROOT_CACHE_DIR, "go");
 pub const BUN_CACHE_DIR: &str = concatcp!(ROOT_CACHE_DIR, "bun");
 pub const HUB_CACHE_DIR: &str = concatcp!(ROOT_CACHE_DIR, "hub");
 pub const GO_BIN_CACHE_DIR: &str = concatcp!(ROOT_CACHE_DIR, "gobin");
+
 
 pub const TAR_PIP_TMP_CACHE_DIR: &str = concatcp!(ROOT_TMP_CACHE_DIR, "tar/pip");
 pub const DENO_TMP_CACHE_DIR: &str = concatcp!(ROOT_TMP_CACHE_DIR, "deno");
 pub const BUN_TMP_CACHE_DIR: &str = concatcp!(ROOT_TMP_CACHE_DIR, "bun");
 pub const GO_TMP_CACHE_DIR: &str = concatcp!(ROOT_TMP_CACHE_DIR, "go");
 pub const HUB_TMP_CACHE_DIR: &str = concatcp!(ROOT_TMP_CACHE_DIR, "hub");
-
+pub const DENO_TMP_CACHE_DIR_DEPS: &str = concatcp!(ROOT_TMP_CACHE_DIR, "deno/deps");
+pub const DENO_TMP_CACHE_DIR_NPM: &str = concatcp!(ROOT_TMP_CACHE_DIR, "deno/npm");
 const NUM_SECS_PING: u64 = 5;
 
 
@@ -261,6 +264,8 @@ lazy_static::lazy_static! {
         .and_then(|x| x.parse::<u64>().ok());
 
     pub static ref CAN_PULL: Arc<RwLock<()>> = Arc::new(RwLock::new(()));
+
+    pub static ref ANSI_ESCAPE_RE: Regex = Regex::new(r"\x1b\[[0-9;]*m").unwrap();
 }
 
 //only matter if CLOUD_HOSTED
@@ -308,12 +313,12 @@ pub async fn run_worker<R: rsmq_async::RsmqConnection + Send + Sync + Clone + 's
     worker_instance: &str,
     worker_name: String,
     i_worker: u64,
-    num_workers: u32,
+    _num_workers: u32,
     ip: &str,
     mut rx: tokio::sync::broadcast::Receiver<()>,
     base_internal_url: &str,
     rsmq: Option<R>,
-    sync_barrier: Arc<RwLock<Option<Barrier>>>,
+    _sync_barrier: Arc<RwLock<Option<Barrier>>>,
 ) {
     #[cfg(not(feature = "enterprise"))]
     if !*DISABLE_NSJAIL {
@@ -460,14 +465,15 @@ pub async fn run_worker<R: rsmq_async::RsmqConnection + Send + Sync + Clone + 's
     if i_worker == 1 {
         if let Some(ref s) = S3_CACHE_BUCKET.clone() {
             let bucket = s.to_string();
-            let copy_to_bucket_tx2 = _copy_to_bucket_tx.clone();
             let worker_name2 = worker_name.clone();
 
+            //piptars can be fetched in background
             handles.push(tokio::task::spawn(async move {
-                tracing::info!(worker = %worker_name2, "Started initial sync in background");
-                join!(copy_denogo_cache_from_bucket_as_tar(&bucket), copy_all_piptars_from_bucket(&bucket));
-                let _ = copy_to_bucket_tx2.send(()).await;
+                tracing::info!(worker = %worker_name2, "Started initial piptar sync in background");
+                copy_all_piptars_from_bucket(&bucket).await;
             }));
+            //denogocache.tar need to be fetched in foreground, block workers until they fetched it
+            copy_denogo_cache_from_bucket_as_tar(s).await;
         }
     }
 
@@ -529,9 +535,9 @@ pub async fn run_worker<R: rsmq_async::RsmqConnection + Send + Sync + Clone + 's
 
                     tracing::info!("Started syncing cache");
                     last_sync = Instant::now();
-                    if num_workers > 1 {
-                        create_barrier_for_all_workers(num_workers, sync_barrier.clone()).await;
-                    }
+                    // if num_workers > 1 {
+                    //     create_barrier_for_all_workers(num_workers, sync_barrier.clone()).await;
+                    // }
                     if let Err(e) = copy_cache_to_tmp_cache().await {
                         tracing::error!("failed to copy cache to tmp cache: {}", e);
                     } else {
@@ -543,25 +549,26 @@ pub async fn run_worker<R: rsmq_async::RsmqConnection + Send + Sync + Clone + 's
                             }
                         }));
                     }
+                    
                 }
             }
 
-            // The barrier is to avoid the sync to bucket syncing partial folders
-            #[cfg(feature = "enterprise")]
-            if num_workers > 1 && S3_CACHE_BUCKET.is_some()  {
-                let read_barrier = sync_barrier.read().await;
-                if let Some(b) = read_barrier.as_ref() {
-                    tracing::debug!("worker #{i_worker} waiting for barrier");
-                    b.wait().await;
-                    tracing::debug!("worker #{i_worker} done waiting for barrier");
-                    drop(read_barrier);
-                    // wait for barrier to be reset
-                    let _ = CAN_PULL.read().await;
-                    tracing::debug!("worker #{i_worker} done waiting for lock");
-                } else {
-                    tracing::debug!("worker #{i_worker} no barrier");
-                };
-            }
+            // // The barrier is to avoid the sync to bucket syncing partial folders
+            // #[cfg(feature = "enterprise")]
+            // if num_workers > 1 && S3_CACHE_BUCKET.is_some()  {
+            //     let read_barrier = sync_barrier.read().await;
+            //     if let Some(b) = read_barrier.as_ref() {
+            //         tracing::debug!("worker #{i_worker} waiting for barrier");
+            //         b.wait().await;
+            //         tracing::debug!("worker #{i_worker} done waiting for barrier");
+            //         drop(read_barrier);
+            //         // wait for barrier to be reset
+            //         let _ = CAN_PULL.read().await;
+            //         tracing::debug!("worker #{i_worker} done waiting for lock");
+            //     } else {
+            //         tracing::debug!("worker #{i_worker} no barrier");
+            //     };
+            // }
             
             let (do_break, next_job) = if first_run { 
                 (false, Ok(Some(QueuedJob::default())))
@@ -594,9 +601,9 @@ pub async fn run_worker<R: rsmq_async::RsmqConnection + Send + Sync + Clone + 's
                         _ = copy_to_bucket_rx.recv() => {
                             tracing::debug!("can_pull lock start");
                             let _lock = CAN_PULL.write().await;
-                            if num_workers > 1 {
-                                create_barrier_for_all_workers(num_workers, sync_barrier.clone()).await;
-                            }
+                            // if num_workers > 1 {
+                            //     create_barrier_for_all_workers(num_workers, sync_barrier.clone()).await;
+                            // }
                             //Arc::new(tokio::sync::Barrier::new(num_workers as usize + 1));
                             #[cfg(feature = "enterprise")]
                             if let Err(e) = copy_tmp_cache_to_cache().await {
@@ -787,21 +794,21 @@ pub async fn run_worker<R: rsmq_async::RsmqConnection + Send + Sync + Clone + 's
 
 }
 
-pub async fn create_barrier_for_all_workers(num_workers: u32, sync_barrier: Arc<RwLock<Option<tokio::sync::Barrier>>>) {
-    tracing::debug!("acquiring write lock");
-    let mut barrier = sync_barrier.write().await;
-    *barrier = Some(tokio::sync::Barrier::new(num_workers as usize));
-    drop(barrier);
-    tracing::debug!("dropped write lock");
-    if let Some(b) = sync_barrier.read().await.as_ref() {
-        tracing::debug!("leader worker waiting for barrier");
-        b.wait().await;
-        tracing::debug!("leader worker done waiting for barrier");
-    };
-    let mut barrier = sync_barrier.write().await;
-    *barrier = None;
-    tracing::debug!("leader worker done waiting for");
-}
+// pub async fn create_barrier_for_all_workers(num_workers: u32, sync_barrier: Arc<RwLock<Option<tokio::sync::Barrier>>>) {
+//     tracing::debug!("acquiring write lock");
+//     let mut barrier = sync_barrier.write().await;
+//     *barrier = Some(tokio::sync::Barrier::new(num_workers as usize));
+//     drop(barrier);
+//     tracing::debug!("dropped write lock");
+//     if let Some(b) = sync_barrier.read().await.as_ref() {
+//         tracing::debug!("leader worker waiting for barrier");
+//         b.wait().await;
+//         tracing::debug!("leader worker done waiting for barrier");
+//     };
+//     let mut barrier = sync_barrier.write().await;
+//     *barrier = None;
+//     tracing::debug!("leader worker done waiting for");
+// }
 pub async fn handle_job_error<R: rsmq_async::RsmqConnection + Send + Sync + Clone>(
     db: &Pool<Postgres>,
     client: &AuthedClient,
@@ -908,7 +915,7 @@ async fn insert_initial_ping(
 
 
 fn extract_error_value(log_lines: &str, i: i32) -> serde_json::Value {
-    return json!({"message": format!("ExitCode: {i}, last log lines:\n{}", log_lines.to_string().trim().to_string()), "name": "ExecutionErr"});
+    return json!({"message": format!("ExitCode: {i}, last log lines:\n{}", ANSI_ESCAPE_RE.replace_all(log_lines.trim(), "").to_string()), "name": "ExecutionErr"});
 }
 
 #[derive(Debug, Clone)]
@@ -1633,7 +1640,7 @@ async fn handle_bash_job(
     Ok(serde_json::json!(logs
         .lines()
         .last()
-        .map(|x| x.to_string())
+        .map(|x| ANSI_ESCAPE_RE.replace_all(x, "").to_string())
         .unwrap_or_else(String::new)))
 }
 
@@ -1672,7 +1679,10 @@ async fn handle_powershell_job(
         .collect::<Vec<(String, String)>>();
     let pwsh_args = args_owned.iter().map(|(n, v)| format!("--{n} {v}")).join(" ");
 
-    let content = content.replace('$', r"\$"); // escape powershell variables
+    let content = content
+        .replace('$', r"\$") // escape powershell variables
+        .replace("`", r"\`"); // escape powershell backticks
+    
     write_file(job_dir, "main.sh", &format!("set -e\ncat > script.ps1 << EOF\n{content}\nEOF\npwsh -File script.ps1 {pwsh_args}\necho \"\"\nsleep 0.02")).await?;
     let token = client.get_token().await;
     let mut reserved_variables = get_reserved_variables(job, &token, db).await?;
@@ -1714,12 +1724,12 @@ async fn handle_powershell_job(
             .stderr(Stdio::piped())
             .spawn()?
     };
-    handle_child(&job.id, db, logs,  child, !*DISABLE_NSJAIL, worker_name, &job.workspace_id, "bash run", job.timeout).await?;
+    handle_child(&job.id, db, logs,  child, !*DISABLE_NSJAIL, worker_name, &job.workspace_id, "bash/powershell run", job.timeout).await?;
     //for now bash jobs have an empty result object
     Ok(serde_json::json!(logs
         .lines()
         .last()
-        .map(|x| x.to_string())
+        .map(|x| ANSI_ESCAPE_RE.replace_all(x, "").to_string())
         .unwrap_or_else(String::new)))
 }
 
@@ -2166,10 +2176,10 @@ async fn capture_dependency_job(
     match job_language {
         ScriptLang::Python3 => {
             create_dependencies_dir(job_dir).await;
-            let req = pip_compile(job_id, job_raw_code, logs, job_dir, db, worker_name, w_id).await;
+            let req: std::result::Result<String, Error> = pip_compile(job_id, job_raw_code, logs, job_dir, db, worker_name, w_id).await;
             // install the dependencies to pre-fill the cache
             if let Ok(req) = req.as_ref() {
-                handle_python_reqs(
+                let r = handle_python_reqs(
                     req
                         .split("\n")
                         .filter(|x| !x.starts_with("--"))
@@ -2182,7 +2192,11 @@ async fn capture_dependency_job(
                     job_dir,
                     worker_dir,
                 )
-                .await?;
+                .await;
+                
+                if let Err(e) = r {
+                    tracing::error!("Failed to install python dependencies to prefill the cache: {:?} \n{}", e, logs);
+                }
             }
             req
         }