+++ frontend/src/lib/components/ScriptBuilder.svelte
@@ -61,21 +61,22 @@
 
 	const langs: [string, SupportedLanguage][] = [
 		['TypeScript (Deno)', Script.language.DENO],
-		['Python', Script.language.PYTHON3]
+		['Python', Script.language.PYTHON3],
+		['TypeScript (Bun)', Script.language.BUN]
 	]
 	if (SCRIPT_SHOW_BASH) {
 		langs.push(['Bash', Script.language.BASH])
 	}
+	if (SCRIPT_SHOW_GO) {
+		langs.push(['Go', Script.language.GO])
+	}
+	langs.push(['REST', Script.language.NATIVETS])
 	langs.push(['PostgreSQL', Script.language.POSTGRESQL])
 	langs.push(['MySQL', Script.language.MYSQL])
 	langs.push(['BigQuery', Script.language.BIGQUERY])
 	langs.push(['Snowflake', Script.language.SNOWFLAKE])
 	langs.push(['GraphQL', Script.language.GRAPHQL])
 	langs.push(['PowerShell', Script.language.POWERSHELL])
-	if (SCRIPT_SHOW_GO) {
-		langs.push(['Go', Script.language.GO])
-	}
-	langs.push(['REST', Script.language.NATIVETS])
 	const scriptKindOptions: {
 		value: Script.kind
 		title: string
@@ -365,23 +366,6 @@
 					}}
 				>
 					<LanguageIcon lang="docker" /><span class="ml-2 py-2">Docker</span>
-				</Button>
-				<Button
-					size="xs"
-					variant="border"
-					color={script.language == 'bun' ? 'blue' : 'light'}
-					btnClasses={script.language == 'bun'
-						? '!border-2 !bg-blue-50/75 dark:!bg-frost-900/75'
-						: 'm-[1px]'}
-					disabled={lockedLanguage}
-					on:click={() => {
-						initContent(Script.language.BUN, script.kind, template)
-						script.language = Script.language.BUN
-					}}
-				>
-					<LanguageIcon lang={Script.language.BUN} /><span class="ml-2 py-2 truncate">
-						TypeScript (Bun, experimental)
-					</span>
 				</Button>
 
 				<!-- <Button